requirements.txt
elasticsearch
requests
pandas
openpyxl
exrex
lxml
BeautifulSoup4
python-docx
html5lib
